+++ tests/PathOfExile.GameClient.Monitor.Tests/Given_DefaultNotificationMatching/When/WhenKitavasAfflictionNotificationMatch.cs
@@ -3,21 +3,20 @@ using PathOfExile.GameClient.Monitor.Notifications;
 
 namespace PathOfExile.GameClient.Monitor.Tests.Given_DefaultNotificationMatching.When
 {
-    public class WhenAfkNotificationMatch : When
+    public class WhenKitavasAfflictionNotificationMatch : When
     {
-        [TestCase(": AFK mode is now ON. Autoreply \"This player is AFK.\"", true)]
-        [TestCase(": AFK mode is now ON.", true)]
-        [TestCase(": AFK mode is now OFF.", false)]
-        public void And_IsMatch(string text, bool isAfk)
+        [TestCase(": You have been permanently weakened by Kitava's cruel affliction. You now have -30% to all Resistances.", "cruel", 30)]
+        [TestCase(": You have been permanently weakened by Kitava's merciless affliction. You now have -60% to all Resistances.", "merciless", 60)]
+        public void And_IsMatch(string text, string difficulty, int penalty)
         {
-            With_AfkNotificationMatch();
+            With_KitavasAfflictionNotificationMatch();
 
             NotificationText = $"{DefaultLogMetadataText} {text}.";
 
             WhenIsMatch();
 
             Then_IsMatch();
-            Then_Notification(new AfkNotification(isAfk, DefaultLogMetadata));
+            Then_Notification(new KitavasAfflictionNotification(difficulty, penalty, DefaultLogMetadata));
         }
 
         [TestCase("AFK mode is now OFF.", Description = "Invalid format - no ':'")]
@@ -25,7 +24,7 @@ namespace PathOfExile.GameClient.Monitor.Tests.Given_DefaultNotificationMatching
         [TestCase("Connect time to instance server was 32ms")]
         public void And_IsNotMatch(string messageText)
         {
-            With_AfkNotificationMatch();
+            With_KitavasAfflictionNotificationMatch();
 
             NotificationText = $"{DefaultLogMetadataText} {messageText}";